+++ json/channels.json
@@ -1,7 +1,7 @@
 {
-  "29b7f628-3d50-40ae-8fe3-4789f0886a00" : {
-    "id" : "29b7f628-3d50-40ae-8fe3-4789f0886a00",
-    "createdAt" : 1737599247292,
+  "e8671ec5-5555-4c80-835c-642788d4be12" : {
+    "id" : "e8671ec5-5555-4c80-835c-642788d4be12",
+    "createdAt" : 1737766782118,
     "updatedAt" : null,
     "name" : "코드잇 공지 채널",
     "description" : "공지 채널입니다",
