+++ src/main/java/com/sprint/mission/discodeit/factory/service/JCFServiceFactory.java
@@ -1,4 +1,4 @@
-package com.sprint.mission.discodeit.factory;
+package com.sprint.mission.discodeit.factory.service;
 
 import com.sprint.mission.discodeit.service.ChannelService;
 import com.sprint.mission.discodeit.service.MessageService;
@@ -7,7 +7,7 @@ import com.sprint.mission.discodeit.service.jcf.JCFChannelService;
 import com.sprint.mission.discodeit.service.jcf.JCFMessageService;
 import com.sprint.mission.discodeit.service.jcf.JCFUserService;
 
-public class JCFServiceFactory implements ServiceFactory{
+public class JCFServiceFactory implements ServiceFactory {
     private static JCFServiceFactory instance;
     private final UserService userService;
     private final MessageService messageService;
