+++ src/main/java/com/sprint/mission/discodeit/service/file/FileUserService.java
@@ -30,11 +30,11 @@ public class FileUserService implements UserService {
 
     @Override
     public User create(String name, String phone, String password) {
-        if (!ValidPass.isValidPassword(password)) {
+        if (!User.isValidPassword(password)) {
             throw new ServiceException(ErrorCode.INVALID_PASSWORD);
         }
 
-        if (!ValidPhone.isValidPhone(phone)) {
+        if (!User.isValidPhone(phone)) {
             throw new ServiceException(ErrorCode.INVALID_WRITER);
         }
         if (users.values().stream()
